+++ test/Compute.NextInterval.t.sol
@@ -46,22 +46,13 @@ contract ComputeNextIntervalPrepareTest is ComputeTest {
         // --- Assertions for the next interval preparation ---
         // 6. Deliver compute for interval 1, which triggers preparation for interval 2
         vm.prank(address(BOB));
-        BOB.reportComputeResult(
-            1,
-            MOCK_INPUT,
-            MOCK_OUTPUT,
-            MOCK_PROOF,
-            commitmentData1,
-            bobWallet
-        );
+        BOB.reportComputeResult(1, MOCK_INPUT, MOCK_OUTPUT, MOCK_PROOF, commitmentData1, bobWallet);
 
         // Expect a new request to be started for interval 2
         bytes32 requestId2 = keccak256(abi.encodePacked(subId, uint32(2)));
         // Expect funds to be locked for the second request
         vm.expectEmit(true, true, false, false, address(Wallet(payable(aliceWallet))));
-        emit Wallet.RequestLocked(
-            requestId2, address(SUBSCRIPTION), address(TOKEN), feeAmount * redundancy, redundancy
-        );
+        emit Wallet.RequestLocked(requestId2, address(SUBSCRIPTION), address(TOKEN), feeAmount * redundancy, redundancy);
         BOB.prepareNextInterval(subId, 2, address(BOB));
         // 7. Final assertions on wallet state
         assertEq(
@@ -81,14 +72,7 @@ contract ComputeNextIntervalPrepareTest is ComputeTest {
         Wallet(payable(consumerWallet)).approve(address(CALLBACK), address(TOKEN), feeAmount);
 
         (, Commitment memory commitment1) = CALLBACK.createMockRequest( //
-            MOCK_CONTAINER_ID,
-            MOCK_CONTAINER_INPUTS,
-            1,
-            address(TOKEN),
-            feeAmount,
-            consumerWallet,
-            NO_VERIFIER
-        );
+        MOCK_CONTAINER_ID, MOCK_CONTAINER_INPUTS, 1, address(TOKEN), feeAmount, consumerWallet, NO_VERIFIER);
 
         // 2. Deliver the compute for the final (and only) interval.
         bytes memory commitmentData1 = abi.encode(commitment1);
@@ -105,7 +89,7 @@ contract ComputeNextIntervalPrepareTest is ComputeTest {
         bytes32 requestStartSelector = Router.RequestStart.selector;
         bytes32 requestLockedSelector = Wallet.RequestLocked.selector;
 
-        for (uint i = 0; i < logs.length; i++) {
+        for (uint256 i = 0; i < logs.length; i++) {
             assertNotEq(logs[i].topics[0], requestStartSelector, "Should not emit RequestStart");
             assertNotEq(logs[i].topics[0], requestLockedSelector, "Should not emit RequestLocked");
         }
@@ -128,16 +112,7 @@ contract ComputeNextIntervalPrepareTest is ComputeTest {
 
         // 4. Create subscription and first request
         (uint64 subId, Commitment memory commitment1) = SUBSCRIPTION.createMockSubscription( //
-            MOCK_CONTAINER_ID,
-            2,
-            1 minutes,
-            redundancy,
-            false,
-            address(TOKEN),
-            feeAmount,
-            consumerWallet,
-            NO_VERIFIER
-        );
+        MOCK_CONTAINER_ID, 2, 1 minutes, redundancy, false, address(TOKEN), feeAmount, consumerWallet, NO_VERIFIER);
         // 5. Deliver compute for the first interval
         vm.warp(block.timestamp + 1 minutes);
         bytes memory commitmentData1 = abi.encode(commitment1);
@@ -149,7 +124,9 @@ contract ComputeNextIntervalPrepareTest is ComputeTest {
         BOB.prepareNextInterval(subId, 2, address(BOB));
         // 7. Assert that no funds were locked for the next interval.
         bytes32 requestId2 = keccak256(abi.encodePacked(subId, uint32(2)));
-        assertEq(Wallet(payable(consumerWallet)).lockedOfRequest(requestId2), 0, "Should not lock funds for next interval");
+        assertEq(
+            Wallet(payable(consumerWallet)).lockedOfRequest(requestId2), 0, "Should not lock funds for next interval"
+        );
     }
 
     function test_DoesNotPrepareNextInterval_When_InsufficientAllowance() public {
@@ -169,16 +146,7 @@ contract ComputeNextIntervalPrepareTest is ComputeTest {
 
         // 4. Create subscription and first request
         (uint64 subId, Commitment memory commitment1) = SUBSCRIPTION.createMockSubscription( //
-            MOCK_CONTAINER_ID,
-            2,
-            1 minutes,
-            redundancy,
-            false,
-            address(TOKEN),
-            feeAmount,
-            consumerWallet,
-            NO_VERIFIER
-        );
+        MOCK_CONTAINER_ID, 2, 1 minutes, redundancy, false, address(TOKEN), feeAmount, consumerWallet, NO_VERIFIER);
 
         // 5. Deliver compute for the first interval
         vm.warp(block.timestamp + 1 minutes);
@@ -212,13 +180,7 @@ contract ComputeNextIntervalPrepareTest is ComputeTest {
         // Fund protocol wallet with ETH
         vm.deal(protocolWallet, 10 ether);
         DeployUtils.updateBillingConfig(
-            COORDINATOR,
-            1 weeks,
-            protocolWallet,
-            MOCK_PROTOCOL_FEE,
-            0,
-            expectedTickFee,
-            address(0)
+            COORDINATOR, 1 weeks, protocolWallet, MOCK_PROTOCOL_FEE, 0, expectedTickFee, address(0)
         );
 
         // The protocol wallet is its own spender for tick fees.
@@ -235,30 +197,14 @@ contract ComputeNextIntervalPrepareTest is ComputeTest {
 
         // Create the subscription
         (uint64 subId, Commitment memory commitment) = SUBSCRIPTION.createMockSubscription( //
-            MOCK_CONTAINER_ID,
-            2,
-            1 minutes,
-            redundancy,
-            false,
-            ZERO_ADDRESS,
-            feeAmount,
-            consumerWallet,
-            NO_VERIFIER
-        );
+        MOCK_CONTAINER_ID, 2, 1 minutes, redundancy, false, ZERO_ADDRESS, feeAmount, consumerWallet, NO_VERIFIER);
         assertEq(subId, 1);
 
         // 2. Act: Deliver compute for the first interval, which triggers preparation for the second.
         vm.warp(block.timestamp + 1 minutes);
         bytes memory commitmentData1 = abi.encode(commitment);
         vm.prank(address(BOB));
-        BOB.reportComputeResult(
-            1,
-            MOCK_INPUT,
-            MOCK_OUTPUT,
-            MOCK_PROOF,
-            commitmentData1,
-            nodeWallet
-        );
+        BOB.reportComputeResult(1, MOCK_INPUT, MOCK_OUTPUT, MOCK_PROOF, commitmentData1, nodeWallet);
         BOB.prepareNextInterval(commitment.subscriptionId, 2, nodeWallet);
         // 3. Assert: Check if the node wallet received the tick fee.
         uint256 computeFee = 0.8978 ether;
